requirements.txt
google-play-scraper==1.2.3
requests==2.31.0
python-dateutil==2.8.2
google-generativeai==0.3.0
